Cleaning/modernizing #20
More improvements: QStringLiterals and removing redundant conditions.

+++ Filters/Filter.h
@@ -24,7 +24,7 @@ public Q_SLOTS:
     virtual void setChecked(bool checked);
 
 protected:
-    long long column_;
+    int64_t column_;
 };
 
 #endif // FILTER_H
